+++ vdmj/src/main/java/com/fujitsu/vdmj/values/SeqValue.java
@@ -120,7 +120,7 @@ public class SeqValue extends Value
 
 	public Value get(Value arg, Context ctxt) throws ValueException
 	{
-		int i = (int)arg.nat1Value(ctxt);
+		int i = (int)arg.nat1Value(ctxt).intValue();
 
 		if (i < 1 || i > values.size())
 		{
